+++ processor/src/test/resources/input/com/example/verifiable/DisableVerifyModel.java
@@ -1,0 +1,24 @@
+package com.example.verifiable;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import arez.annotations.Observable;
+import arez.annotations.Reference;
+import arez.annotations.ReferenceId;
+
+@ArezComponent( allowEmpty = true, verify = Feature.DISABLE )
+abstract class DisableVerifyModel
+{
+  @Reference
+  abstract MyEntity getMyEntity();
+
+  @ReferenceId
+  @Observable
+  abstract int getMyEntityId();
+
+  abstract void setMyEntityId( int id );
+
+  static class MyEntity
+  {
+  }
+}