+++ processor/src/test/resources/input/com/example/verifiable/EnableVerifyModel.java
@@ -1,0 +1,9 @@
+package com.example.verifiable;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+
+@ArezComponent( verify = Feature.ENABLE, allowEmpty = true )
+abstract class EnableVerifyModel
+{
+}