+++ CHANGELOG.md
@@ -31,6 +31,8 @@
   unused reference support code.
 * **\[core\]** Expose a per-context `arez.Locator` via `ArezContext.locator()`. This implementation delegates to
   instances of `arez.Locator` registered via `ArezContext.registerLocator(...)`.
+* **\[core\]** Add the `verify` parameter to the `@ArezComponent` annotation that controls whether the generated
+  component will implement the `arez.component.Verifiable` interface and perform verification at runtime.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)